backend/requirements.txt
Django==5.2.6
asgiref==3.8.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.8.0
environs[django]==14.3.0
psycopg2-binary==2.9.10
django-filter==25.1
drf-spectacular==0.28.0
Pillow==11.3.0
django-storages[s3]==1.14.6